examples/pdf_grobid_edge_tts/requirements.txt
python-dotenv
spacy
en_core_web_sm
beautifulsoup4
edge-tts